+++ src/api/controllers/users-controller.ts
@@ -1,5 +1,5 @@
-import { RequestCustom, ResponseCustom } from '../interfaces';
-import { usersService } from 'src/services';
+import { RequestCustom, ResponseCustom } from '../../interfaces';
+import { usersService } from 'src/api/services';
 
 const { getAllUsers, getUserById, createUser, updateUser, deleteUser } = usersService;
