+++ src/constants/index.ts
@@ -1,3 +1,10 @@
+import { config } from 'dotenv';
+
+config();
+
+export const defaultPort = 5000;
+export const PORT = process.env.PORT || defaultPort;
+export const baseURL = `http://localhost:${PORT}`;
 export const usersEndpoint = '/api/users';
 
 export const enum StatusCodes {
