+++ src/utils/searchPattern.ts
